requirements.txt
fastapi
uvicorn
pydantic
groq
faiss-cpu
sentence-transformers
beautifulsoup4
requests
httpx
pytest
pytest-asyncio
python-dotenv
numpy
scikit-learn
